requirements.txt
flask==3.0.0
tensorflow==2.16.1
opencv-python==4.8.1.78
numpy>=1.24.3
Pillow>=10.0.0
sqlalchemy>=2.0.0
flask-sqlalchemy>=3.1.1
Werkzeug>=3.0.0